% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdrawr-package.R
\docType{package}
\name{distdrawr-package}
\alias{distdrawr-package}
\title{The distdrawr package}
\description{
Tools for downloading data from the FLORKART-database on the
distribution of vascular plants in Germany in a convenient way.
}
\references{
Datenbank FLORKART der Floristischen Kartierung Deutschlands,
Stand 2013, Bundesamt fuer Naturschutz (BfN) und Netzwerk Phytodiversitaet
Deutschland (NetPhyD): \url{http://www.floraweb.de}
}
\seealso{
\code{\link{get.dist} \link{check_species}}
}
\author{
Nicolas Friess
}
